refactor: ScrapService와 SopticleService 인터페이스 분리
